+++ qsprpred/data/utils/descriptorsets.py
@@ -6,14 +6,17 @@ To add a new descriptor or fingerprint calculator:
 """
 
 from abc import ABC, abstractmethod
+from typing import List, Union
 
 import mordred
 import numpy as np
+import pandas as pd
 from mordred import descriptors as mordreddescriptors
 from qsprpred.data.utils.descriptor_utils import fingerprints
 from qsprpred.data.utils.descriptor_utils.drugexproperties import Property
 from qsprpred.data.utils.descriptor_utils.rdkitdescriptors import RDKit_desc
 from rdkit import Chem, DataStructs
+from rdkit.Chem import AllChem, Mol
 
 
 class DescriptorSet(ABC):
@@ -23,17 +26,33 @@ class DescriptorSet(ABC):
     """
 
     @abstractmethod
-    def __call__(self, mol):
+    def __call__(self, mols: List[Union[str, Mol]]):
         """
         Calculate the descriptor for a molecule.
 
         Args:
-            mol: smiles or rdkit molecule
+            mols: list of molecules (SMILES `str` or RDKit Mol)
 
         Returns:
-            a `list` of descriptor values
+            an array or data frame of descriptor values of shape (n_mols, n_descriptors)
         """
         pass
+
+    def iterMols(self, mols: List[Union[str, Mol]], to_list=False):
+        """
+        Calculate the descriptor for a molecule.
+
+        Args:
+            mols: list of molecules (SMILES `str` or RDKit Mol)
+
+        Returns:
+            an array or data frame of descriptor values of shape (n_mols, n_descriptors)
+        """
+
+        ret = (Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol for mol in mols)
+        if to_list:
+            ret = list(ret)
+        return ret
 
     @property
     @abstractmethod
@@ -85,18 +104,20 @@ class FingerprintSet(DescriptorSet):
 
         self._keepindices = None
 
-    def __call__(self, mol):
-        """Calculate the fingerprint for a molecule."""
-        convertMol = Chem.MolFromSmiles
+    def __call__(self, mols):
+        """Calculate the fingerprint for a list of molecules."""
         convertFP = DataStructs.ConvertToNumpyArray
 
-        mol = convertMol(mol) if isinstance(mol, str) else mol
-        fp = self.get_fingerprint(mol)
-        ret = np.zeros(len(fp))
-        convertFP(fp, ret)
-        if self.keepindices:
-            ret = ret[self.keepindices]
-        return list(ret)
+        ret = np.zeros((len(mols), self.get_len()))
+        for idx, mol in enumerate(self.iterMols(mols)):
+            fp = self.get_fingerprint(mol)
+            np_fp = np.zeros(len(fp))
+            convertFP(fp, np_fp)
+            if self.keepindices:
+                np_fp = np_fp[self.keepindices]
+            ret[idx] = np_fp
+
+        return ret
 
     @property
     def keepindices(self):
@@ -120,7 +141,7 @@ class FingerprintSet(DescriptorSet):
 
     def get_len(self):
         """Return the length of the fingerprint."""
-        return len(self.__call__("C"))
+        return len(self.get_fingerprint(Chem.MolFromSmiles("C")))
 
     def __str__(self):
         return f"FingerprintSet"
@@ -181,9 +202,8 @@ class Mordred(DescriptorSet):
         # convert to list of descriptor names if descriptor instances are passed and initiate mordred calulator
         self.descriptors = [str(d) for d in descs]
 
-    def __call__(self, mol):
-        mol = Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol
-        return self._mordred.pandas([mol], quiet=True, nproc=1).iloc[0].to_list()
+    def __call__(self, mols):
+        return self._mordred.pandas(self.iterMols(mols), quiet=True, nproc=1).values
 
     @property
     def is_fp(self):
@@ -234,11 +254,10 @@ class DrugExPhyschem(DescriptorSet):
         self._is_fp = False
         self.props = [x for x in Property(physchem_props).props]
 
-    def __call__(self, mol):
+    def __call__(self, mols):
         """Calculate the DrugEx properties for a molecule."""
         calculator = Property(self.props)
-        mol = Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol
-        return list(calculator.getScores([mol])[0])
+        return calculator.getScores(self.iterMols(mols, to_list=True))
 
     @property
     def is_fp(self):
@@ -276,9 +295,8 @@ class rdkit_descs(DescriptorSet):
         self._descriptors = self._calculator.descriptors
         self.compute_3Drdkit = compute_3Drdkit
 
-    def __call__(self, mol):
-        mol = Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol
-        return list(self._calculator.getScores([mol])[0])
+    def __call__(self, mols):
+        return self._calculator.getScores(self.iterMols(mols, to_list=True))
 
     @property
     def is_fp(self):
@@ -366,6 +384,60 @@ class TanimotoDistances(DescriptorSet):
         return "TanimotoDistances"
 
 
+class PredictorDesc(DescriptorSet):
+    """DescriptorSet that uses a Predictor object to calculate the descriptors for a molecule."""
+
+    def __init__(self, *args, **kwargs):
+        """
+        Initialize the descriptorset with a Predictor object.
+
+        Args:
+            predictor: Predictor object to use for calculating the descriptor
+        """
+        self._args = args
+        self._kwargs = kwargs
+        self._is_fp = False
+        from qsprpred.scorers.predictor import Predictor
+        self._predictor = Predictor.fromFile(*args, **kwargs)
+        self._descriptors = [self._predictor.getKey()]
+
+    def __call__(self, mol):
+        """
+        Calculate the descriptor for a molecule.
+
+        Args:
+            mol: smiles or rdkit molecule
+
+        Returns:
+            a `list` of descriptor values
+        """
+        mol = Chem.MolFromSmiles(mol) if isinstance(mol, str) else mol
+        return list(self._predictor.getScores([mol]))
+
+    @property
+    def is_fp(self):
+        return self._is_fp
+
+    @property
+    def settings(self):
+        """Return args and kwargs used to initialize the descriptorset."""
+        return self._args, self._kwargs
+
+    @property
+    def descriptors(self):
+        return self._descriptors
+
+    @descriptors.setter
+    def descriptors(self, descriptors):
+        self._descriptors = descriptors
+
+    def get_len(self):
+        return 1
+
+    def __str__(self):
+        return "PredictorDesc"
+
+
 class _DescriptorSetRetriever:
     """Based on recipe 8.21 of the book "Python Cookbook".
 
@@ -390,6 +462,9 @@ class _DescriptorSetRetriever:
 
     def get_RDkit(self, *args, **kwargs):
         return rdkit_descs(*args, **kwargs)
+
+    def get_PredictorDesc(self, *args, **kwargs):
+        return PredictorDesc(*args, **kwargs)
 
     def get_TanimotoDistances(self, *args, **kwargs):
         return TanimotoDistances(*args, **kwargs)